tools: introduce flat_array_ops.h
Currently, the file array_ops.h provides helper functions to
allocate 2D, 3D, and N_3D arrays by offering Create2DArray(),
Create3DArray(), and Create_N_3DArray(). Further, the file
array_ops.cpp provides additional functions Create3DArray_v4sf()
and Create_N_3DArray_v4sf() for allocating vectorized arrays.

However, these functions have several disadvantages due to
its non-contiguous memory layout.

To allow the access of an element in a multi-dimensional array
from a pointer using the C syntax array_ptr[x][y][z], the multi-
dimensional array is allocated in the following way: First, it
allocates an 1D array of pointers with size "x", each pointer
is pointed to another dynamically-allocated 1D array of pointers
with size "y", then each pointer within this array is pointed to
another dynamically-allocated 1D array with size "z".

This is essentially an unoptimized form of "iliffe vector", with
significant performance bottleneck.

First, poor memory spatial locality. If two elements are stored
next to each other in memory and shares the same cacheline, accessing
the first element effectively fetches the next element at the same
time, saving valuable memory bandwidth - this is advantageous since
openEMS's FDTD kernel is a linear walk of all points in 3D space.
Similarly, when an element in memory is accessed with a regular
stride size, the CPU can correctly predict and prefetch the next
elements.

However, in the current implementation, only the last dimension
of a multi-dimensional array is stored contiguous in memory,
what's found in between is bookkeeping data used internally by
malloc(), and this is the best case scenario that assumes the
system is not running with fragmented memory due to other memory-
heavy workloads. As a result, a linear walk of the array cannot
take full advantage of memory locality. When we're accessing
array[x][y][Z_MAX] in a linear walk, array[x][y + 1][0] should be
prefetched, but this is not possible if the last dimension of the
array is not contiguous in memory.

Second, using such an iliffe vector means accessing each element
in the array involves multiple levels of pointer dereferencing,
to access a single element, multiple unrelated memory addresses
must be read first, this further increases memory latency and
reduces the effectiveness of the CPU prefetcher.

The most seriously affected array is the N_3D arrays. These arrays
are effectively 4D arrays as N is currently hardcoded to 3. This
kind of arrays are used to store the polarization of the electro-
magnetic field in the X, Y, Z axis at each point (A, B, C) in 3D
space, and they're the most heavily used data structure.

The new template functions in flat_array_ops.h provides a solution
to these problems. It creates a new class named Flat_N_3DArray<T>.
By calling the function Create_Flat_N_3DArray<T>, it allocates a
contiguous 1D array in memory. The operator() of Flat_N_3DArray<T>
is then overloaded, allowing one to access the element transparently.
Furthermore, instead of creating an unnecessary Create_N_3DArray_v4sf()
to handle vector data types, it instead uses a template specialization
for the type f4vector.  Thus, both scalar and vector arrays can be
created using the same template syntax.

Since the the most seriously affected array is the N_3D arrays, only
the function Create_Flat_N_3DArray<T> at this time, and the plan is
to replace all occurrence of Create_N_3DArray() and Create_N_3DArray_v4sf()
completely with Create_Flat_N_3DArray(). In the future, the plan is
to also replace Create3DArray() and Create2DArray() by their "flat"
equivalent too.

To move the existing code Create_N_3DArray() and Create_N_3DArray_v4sf()
to Create_Flat_N_3DArray(), the following change is needed. If the
original code is

    // initialization
    FDTD_FLOAT**** array;
    array = Create_N_3DArray<FDTD_FLOAT>(...);

    // use
    array[n][x][y][z] = 42;

It should be replaced with:

    // initialization
    Flat_N_3DArray<FDTD_FLOAT>* array;
    array = Create_Flat_N_3DArray<FDTD_FLOAT>(...);

    // use
    (*array)(n, x, y, z) = 42;

However, since syntax "(*array)(n, x, y, z)" is ugly and confusing,
it's recommended to keep the pointer to the array in a data structure,
but when an access to the array is needed, the pointer is converted
a referenece at the beginning of a function, in a style similar to:

    // initialization
    Flat_N_3DArray<FDTD_FLOAT>* array_ptr;
    array_ptr = Create_Flat_N_3DArray<FDTD_FLOAT>(...);

    // use
    Flat_N_3DArray<FDTD_FLOAT>& array = *array_ptr;
    array(n, x, y, z) = 42;

Signed-off-by: Yifeng Li <tomli@tomli.me>