FDTD: change FDTD kernel to support tiling.
This commit changes the definitions of UpdateVoltages()
and UpdateCurrents() from:

    UpdateVoltages(unsigned int startX, unsigned int numX);
    UpdateCurrents(unsigned int startX, unsigned int numX);

To:

    virtual void UpdateVoltages(int start[3], int stop[3]);
    virtual void UpdateCurrents(int start[3], int stop[3]);

Instead of running domain decomposition only on the X axis,
now the FDTD kernel can run on a decomposed on the Y axis too.

Due to SIMD complications, the SSE engines don't support tiling
on the Z axis. IF the input Z range is not the full range, the
program terminates.

The semantics of these functions have also been changed, stop[3]
is inclusive, not exclusive. It's the caller's responsibility to
check whether the stop line causes out-of-bound access, it means
the caller has to avoid passing the last line to UpdateCurrents().

Signed-off-by: Yifeng Li <tomli@tomli.me>