engine_extension: new APIs for loop tiling.
Using the existing engine extension APIs, the internal states
of the engine and the voltages and currents of the simulation
domain can be updated in two ways:

1. Updating the entire simulation domain by not passing additional
argument.

2. Asking the extensions to decide where to update by themselves,
based on its "int threadID".

To allow the new engine implementation with spatial and temporal
tiling, this commit introduces a new set of engine extention APIs:

    virtual void InitializeTiling(std::vector<Range3D> tiles);

    DoPreVoltageUpdates(int timestep, unsigned int start[3], unsigned int stop[3]);
    DoPostVoltageUpdates(int timestep, unsigned int start[3], unsigned int stop[3]);
    Apply2Voltages(int timestep, unsigned int start[3], unsigned int stop[3]);
    DoPreCurrentUpdates(int timestep, unsigned int start[3], unsigned int stop[3]);
    DoPostCurrentUpdates(int timestep, unsigned int start[3], unsigned int stop[3]);
    Apply2Current(int timestep, unsigned int start[3], unsigned int stop[3]);

In these APIs, the updated range is strictly contained within "int start[3]"
and "int stop[3]". It's the engine's responsibility to partition the
space into tiles, not the extentions. Furthermore, since the engine may
calculate many timesteps at once in the same tiles, but some extentions
may do something only at a specific timestep, the tiling API also needs
to know the current timestep.

To implement tiling in an extention, the engine should check whether the
current range is relevant, if so, this range is updated. If not, this
range is ignored. The updated range should not depend on the values from
other ranges - "action at a distance" is not allowed.

Further, the InitializeTiling() method is used for passing all tiles
involved to the extentions in advance for initialization, it's only
used by the tiling engine, in analogues to the role of the
SetNumberOfThreads() method used by the standard multi-threading engine.

Signed-off-by: Yifeng Li <tomli@tomli.me>