openems.cpp: use Tiling engine by default (hack).
This commit is a temporary hack to use the spatial/temporal
Tiling engine by default. Since not all features and extentions
are supported, it's not a suitable default. However, the Python
API lacks the ability to select the underlying engine, so the
new Tiling engine is temporary used by default for testing.

Signed-off-by: Yifeng Li <tomli@tomli.me>