Initial Kokkos engine
This is the first prototype of the Kokkos engine for the CPU. Due to
the lack of both loop tiling and operator compression, the performance
is in fact worse than the existing engine. This is only meant for test
during early development, and should not be judged in terms of performance
or correctness.